Add Emscripten WebGPU support

+++ Source/WGPU/ImplWGPU.cpp
@@ -207,14 +207,14 @@ static Result NRI_CALL CreateTexture(Device& device, const TextureDesc& textureD
 
 static void NRI_CALL GetBufferMemoryDesc(const Buffer& buffer, MemoryLocation memoryLocation, MemoryDesc& memoryDesc) {
     const BufferDesc& bufferDesc = ((BufferWGPU&)buffer).GetDesc();
-    memoryDesc = {std::max(bufferDesc.size, 1ull), 1, (MemoryType)memoryLocation, false};
+    memoryDesc = {std::max(bufferDesc.size, uint64_t(1)), 1, (MemoryType)memoryLocation, false};
 }
 
 static void NRI_CALL GetTextureMemoryDesc(const Texture& texture, MemoryLocation memoryLocation, MemoryDesc& memoryDesc) {
     const TextureDesc& textureDesc = ((TextureWGPU&)texture).GetDesc();
     uint64_t size = (uint64_t)textureDesc.width * std::max((Dim_t)1, textureDesc.height) * std::max((Dim_t)1, textureDesc.depth) * std::max((Dim_t)1, textureDesc.layerNum) * std::max((Dim_t)1, textureDesc.mipNum);
 
-    memoryDesc = {std::max(size, 1ull), 1, (MemoryType)memoryLocation, false};
+    memoryDesc = {std::max(size, uint64_t(1)), 1, (MemoryType)memoryLocation, false};
 }
 
 static Result NRI_CALL BindBufferMemory(const BindBufferMemoryDesc* bindBufferMemoryDescs, uint32_t bindBufferMemoryDescNum) {
@@ -234,13 +234,13 @@ static Result NRI_CALL BindTextureMemory(const BindTextureMemoryDesc*, uint32_t)
 }
 
 static void NRI_CALL GetBufferMemoryDesc2(const Device&, const BufferDesc& bufferDesc, MemoryLocation memoryLocation, MemoryDesc& memoryDesc) {
-    memoryDesc = {std::max(bufferDesc.size, 1ull), 1, (MemoryType)memoryLocation, false};
+    memoryDesc = {std::max(bufferDesc.size, uint64_t(1)), 1, (MemoryType)memoryLocation, false};
 }
 
 static void NRI_CALL GetTextureMemoryDesc2(const Device&, const TextureDesc& textureDesc, MemoryLocation memoryLocation, MemoryDesc& memoryDesc) {
     uint64_t size = (uint64_t)textureDesc.width * std::max((Dim_t)1, textureDesc.height) * std::max((Dim_t)1, textureDesc.depth) * std::max((Dim_t)1, textureDesc.layerNum) * std::max((Dim_t)1, textureDesc.mipNum);
 
-    memoryDesc = {std::max(size, 1ull), 1, (MemoryType)memoryLocation, false};
+    memoryDesc = {std::max(size, uint64_t(1)), 1, (MemoryType)memoryLocation, false};
 }
 
 static Result NRI_CALL CreateCommittedBuffer(Device& device, MemoryLocation memoryLocation, float priority, const BufferDesc& bufferDesc, Buffer*& buffer) {
